pyproject.toml
[tool.poetry]
name = "EE"
version = "0.1.0"
description = "Multimodal EE Environment"
authors = ["ANON"]

readme = 'README.md'  # Markdown files are supported
keywords = ['research', 'evaluation', 'classification', 'documetns']

classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Web Environment",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.1"
evaluate = "^0.4.0"
scikit-learn = "^1.2.0"
transformers = "^4.26.0"
pandas = "^1.5.3"
torch = "^1.13.1"
pillow = "^9.4.0"
datasets = "^1.16.0"
pytesseract = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^3.10.1"
pytest-cov = "^2.9.0"
git-lfs = "^2.13.3"
libtesseract-dev = "^4.1.1"
libleptonica-dev = "^1.80.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"
